feat(today events): implement next page only pagination
Due to a framework bug - calling `mutate()` on a resource does not cause
a re render. Infinite scrolling is not possible to implement atm. So
forward only pagination is implemented for now.

Not allowing to go back to the previous page makes implementation
simpler. No more need to store a list of cursors in local storage. The
after cursor can now be handled in memory.