feat: meetup events today (#53)
* feat(backend): category search module

* feat(backend): `/meetup/today`

Get meetup events for today

* fix(backend): dont provide default sha hash

It seems like different request type has a different hash. So it
shouldnt have a default value

* feat(frontend): disable pagination buttons if no next page

* feat(today events): implement next page only pagination

Due to a framework bug - calling `mutate()` on a resource does not cause
a re render. Infinite scrolling is not possible to implement atm. So
forward only pagination is implemented for now.

Not allowing to go back to the previous page makes implementation
simpler. No more need to store a list of cursors in local storage. The
after cursor can now be handled in memory.

* fix(backend): `/meetup/today` query param

* fix(frontend): Make pagination component only forward pagination

* fix(frontend): Meetup component does only next page pagination

No more previous page pagination